Upgrade cutlass to v3.8.0 with commit 833f699 (#131)
* Handle MNK Sm90{Row, Col}Reduction problem shapes (#1803)

* add is_last_tile

* Improve sm90 mixed dtype kernel (#1883)

* Add GMMA shape m64n40k16 (#1864)

* Add all supported GMMA shapes (#1890)

* add maximum support (#1833)

* fix typo (#1853)

* fix by adding public (#1753)

* added mapping for bf16 to torch::kBFloat16 (#1843)


* Fix README (#1658)

* Fix README

* Improve README

---------


* Adjusting code indentation (#1639)

* Include of regular_tile_iterator.h fixed for NVRTC (#1765)

* Include of regular_tile_iterator.h fixed for NVRTC

* More include fixed for NVRTC

* Update gemm_f16n_f16t_f32t_tensor_op_f32_sm80.cu with include "cutlass/gemm/device/gemm_universal.h" (#1569)

fix compile with `cmake .. -DCUTLASS_ENABLE_TESTS=ON -DCUTLASS_TEST_LEVEL=2`

* remove redundant hardcoded packing configs in mixed dtype gemm (#1894)


* fix wrong A/BLayout in MMA_Traits for binary mma and append other MMA_Traits support  (#1856)

* fix wrong A/BLayout in  MMA_Traits<SM80_16x8x256_S32U1U1S32_TN_XORPOPC> and append support for  m8n8k128, m16n8k128  mma.and.popc in MMA_Traits instantiation

* add "print" template for  subbyte_reference<T>

* Add a print for the uint{x}b_t type. (#1871)

* Refactor some GroupedGEMM logic (#1899)

* feat: support kFactor 8 used in mma tensor op tile iterator (#1512)

* Update publications (#1912)

* remove restriction of stride == kernel in nhwc_pooling (#1896)

* fix undefined in device code error (#1880)

* Fix the racing condition of mixed-input gemm when writing the registers (#1931)

* move two warpgroup_wait

* merge main

---------


* Fix `cutlass` python library with cuda `12.6.2.post1` (#1942)

* Fix `cutlass` python library with cuda `12.6.2.post1`

Previously we had this error:
```
  File "/storage/home/cutlass/python/cutlass/backend/operation.py", line 39, in <listcomp>
    _version_splits = [int(x) for x in __version__.split("rc")[0].split(".")]
                       ^^^^^^
ValueError: invalid literal for int() with base 10: 'post1'
```

* Update sm90_utils.py

* Update generator.py

* Update python/cutlass_library/generator.py


* Update python/cutlass_library/sm90_utils.py


---------


* add {uint4, uint2, int2} => {fp16, bf16} conversion (#1966)

* Improve mixed dtype GEMM (#1972)

* update

* fix a typo

* fix a typo that fails the compiling when ElementScale is not the same as MmaType (#1977)

* Fix CuTe README Typo (#1951)

* Fix Typo (#1962)

* 3.6.0 update (#2005)

* 3.6.0 update

* doc and swap stuff

---------


* Update CHANGELOG.md

* Update 0x_gemm_tutorial.md (#1982)

Shouldn't this be BLK_M, BLK_**K**, k

* fix bug: arch/mma_sm60.h Mma<2,2,1> calculate wrong (#1989)

* fix mem fence (#2030)


* Add half->int8 saturate conversion to promise valid range (#1983)

* Add half->int8 saturate conversion to promise valid range

* add gpu only macro

---------


* Add vector-types back to platform.h (#2026)

* Fix typo in library_defaults.py (#2024)

* Fix Typos (#2021)

* Fix Typo

* Fix Typo

* Add Line Break (#2020)

* Blockwise Scaling for FP8 (#1932)

* F8 Blockwise Scaling

* two more NumProducerThreadEvents

---------


* fix assertion in integer_subbytes.h (#1961)

* CUTLASS 3.7 (#2045)

* CUTLASS 3.7

* clean up changelog

---------


* update 3.7 docs (#2051)

* update docs

* update docs

* update docs

* update docs

* update docs

---------


* CUTLASS 3.8 Release (#2059)

* CUTLASS 3.8 Release

* update

* Update README.md

* Revert "Update README.md"

This reverts commit b353e36fe83e0815f99b44e46c0c95494c44726b.

* update

* update

---------


* fix cuda 12.6 issues (#2066)

* fix a readme broken link (#2069)

* Update README.md

* Groupwise scaling along M for FP8 gemm (#2037)

* FP8 groupwise scaling along M

* small updates

---------


* bugfix generic-k code in top-k with softmax (#1993)

* bugfix generic-k code in top-k with softmax

* Update include/cutlass/epilogue/fusion/sm90_visitor_topk_softmax.hpp


* Update examples/61_hopper_gemm_with_topk_and_softmax/61_hopper_gemm_with_topk_and_softmax.cu


---------


* [EVT] Add support for Row/Col broadcast PtrArray (#2033)

* Add group support to EVT row/col broadcast.

* small modifications

---------


* v3.8.0 update (#2082)

* 3.8 update

* fix Markus' name

---------


* [WA] Fix compiling errors

---------

+++ include/cute/util/debug.hpp
@@ -1,5 +1,5 @@
 /***************************************************************************************************
- * Copyright (c) 2023 - 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * Copyright (c) 2023 - 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: BSD-3-Clause
  *
  * Redistribution and use in source and binary forms, with or without
@@ -123,7 +123,7 @@ bool
 block([[maybe_unused]] int bid)
 {
 #if defined(__CUDA_ARCH__)
-  return blockIdx.x + blockIdx.y*gridDim.x + blockIdx.z*gridDim.x*gridDim.y == bid;
+  return blockIdx.x + blockIdx.y*gridDim.x + blockIdx.z*gridDim.x*gridDim.y == static_cast<unsigned int>(bid);
 #elif defined(SYCL_LANGUAGE_VERSION)
   return sycl::ext::oneapi::experimental::this_group<3>().get_group_linear_id() == bid;
 #else
@@ -136,7 +136,7 @@ bool
 thread([[maybe_unused]] int tid, [[maybe_unused]] int bid)
 {
 #if defined(__CUDA_ARCH__)
-  return (threadIdx.x + threadIdx.y*blockDim.x + threadIdx.z*blockDim.x*blockDim.y == tid) && block(bid);
+  return (threadIdx.x + threadIdx.y*blockDim.x + threadIdx.z*blockDim.x*blockDim.y == static_cast<unsigned int>(tid)) && block(bid);
 #elif defined(SYCL_LANGUAGE_VERSION)
   return (sycl::ext::oneapi::experimental::this_group<3>().get_local_linear_id() == tid) && block(bid);
 #else
